- O painel lateral direito foi refeito, com algumas opções removidas dela e sendo delegadas para a nova janela de opções.
- Adicionada uma nova janela de opções contendo as definições de parâmetros gerais do programa. Tal janela pode ser aberta ao clicar no botão com uma imagem de uma engrenagem, na nova barra de ferramentas.

- Adicionadas duas novas opções configuráveis: Escala de tempo e Número de passos por frame. A primeira controla o valor da escala de tempo já usada desde a primeira implementação desse sistema. A segunda controla quantas iterações serão realizadas a cada frame, em outras palavras, ao invés de fazer apenas uma iteração por frame (que antes causava imprecisão na simulação) agora podem ser realizadas varias iterações a cada frame, tornando a simulação muito mais precisa.

- Adicionada uma barra de ferramentas contendo diversas funcionalidades, algumas que foram removidas do painel lateral direito (play e pause por exemplo), além de conter algumas funcionalidades inéditas como criar, abrir e salvar arquivos, além do botão para abrir a janela de opções já mencionada anteriormente.

- Adicionada uma barra de menu contendo por enquanto somente o menu Arquivo que contém os sub-menus de operações com arquivos e uma opção para fechar o programa.

- O arquivo de configuração do programa não salva mais as opções da simulação e nem a configuração de cada corpo, ao invés disso é salvo o nome do arquivo para o qual foi salvo o estado atual da simulação.

- O arquivo de configuração do programa agora salva a posição e dimensões da janela principal, alé de seu estado maximizado.

- Alterada os tipos de diversos dados de float para double na struct Vector2D e classe Body para maior precisão na simulação.

- Realizado pequenos fixes.

+++ Consts.cs
@@ -6,13 +6,13 @@ namespace Orbital
 {
     public static class Consts
     {
-        public const float G = 6.67E-11F;
-        public const float SUN_MASS = 1.989E30F;
-        public const float EARH_MASS = 5.972E24F;
-        public const float MOON_MASS = 7.36E22F;
-        public const float DISTANCE_EARH_SUN = 1.496E11F;
-        public const float DISTANCE_EARH_MOON = 3.844E8F;
-        public const float EARH_SPEED = 3.02E4F;
-        public const float MOON_SPEED = 950;
+        public const double G = 6.67E-11;
+        public const double SUN_MASS = 1.989E30;
+        public const double EARH_MASS = 5.972E24;
+        public const double MOON_MASS = 7.36E22;
+        public const double DISTANCE_EARH_SUN = 1.496E11;
+        public const double DISTANCE_EARH_MOON = 3.844E8;
+        public const double EARH_SPEED = 3.02E4;
+        public const double MOON_SPEED = 950;
     }
 }
